aísla generación de reporte en una clase utilizable por request o consola
